+++ mall-promotion/src/main/resources/application.yml
@@ -11,11 +11,7 @@ spring:
     nacos:
       discovery:
         server-addr: 127.0.0.1:8848
-    alicloud:
-      access-key: LTAI4GGAsG4FgoKpE3Ct1Nyd
-      secret-key: NhCANzAJz0gZHFprQGEF8x559s7fEK
-      oss:
-        endpoint: oss-cn-hangzhou.aliyuncs.com
+
   application:
     name: promotion-service
