+++ mall-tripartite/pom.xml
@@ -20,11 +20,7 @@
 	</properties>
 
 	<dependencies>
-		<dependency>
-			<groupId>com.dangdang.bookmall</groupId>
-			<artifactId>mall-common</artifactId>
-			<version>1.0-SNAPSHOT</version>
-		</dependency>
+
 		<!--oss-->
 		<dependency>
 			<groupId>com.alibaba.cloud</groupId>
@@ -50,6 +46,13 @@
 					<artifactId>junit-vintage-engine</artifactId>
 				</exclusion>
 			</exclusions>
+		</dependency>
+
+
+		<!--        nacos 服务注册发现-->
+		<dependency>
+			<groupId>com.alibaba.cloud</groupId>
+			<artifactId>spring-cloud-starter-alibaba-nacos-discovery</artifactId>
 		</dependency>
 	</dependencies>
