+++ mall-tripartite/src/main/java/com/dangdang/bookmall/tripartite/MallTripartiteApplication.java
@@ -2,7 +2,10 @@ package com.dangdang.bookmall.tripartite;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
+@EnableDiscoveryClient
 @SpringBootApplication
 public class MallTripartiteApplication {
