+++ mall-tripartite/src/main/resources/application.yml
@@ -1,0 +1,10 @@
+server:
+  port: 20001
+
+spring:
+  application:
+    name: tripartite-service
+  cloud:
+    nacos:
+      discovery:
+        server-addr: 127.0.0.1:8848
